third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEOffsetElement-svgdom-dx-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'dx' property of the SVGFEOffsetElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS offsetElement.dx.baseVal is 50
PASS offsetElement.dx.baseVal is 0
PASS successfullyParsed is true

TEST COMPLETE
